+++ CHANGELOG.md
@@ -3,6 +3,11 @@ Change Log
 
 All notable changes to this project will be documented in this file.
 
+## [0.2.1]
+
+- error messages can be overridden by option in constructor
+- format hints can be overridden by option in constructor
+
 ## [0.2] - 2019-11-19
 
 - Added patterns for all EU countries